Updated Tiles & Vegetables, Partially Implemented New Tiles and Salads
The "Table" tile can be used by the player to easily pick up and move vegetables or salads to pick up later.

The "VeggieBox" tile is where the player goes in order to grab new vegetables. Vegetables will spawn endlessly from them.

The "CuttingBoard" tile is partially implemented, but some functionality is missing at this point, such as the ability to add more than a single ingredient to a salad. I will be refining the code and adding the needed functionality in my next commit.